+++ src/s3m/stream.rs
@@ -137,7 +137,7 @@ async fn fold_fn<'a>(
             // if buffer size > buf_size create another buffer and upload the previous one
             if buffer.len() + bytes.len() >= buf_size {
                 count += bytes.len();
-                pb.set_message(&bytesize::to_string(count as u64, true));
+                pb.set_message(bytesize::to_string(count as u64, true));
 
                 let mut new_buf = BytesMut::with_capacity(buf_size);
                 new_buf.put(bytes);
@@ -162,7 +162,7 @@ async fn fold_fn<'a>(
                 })
             } else {
                 count += bytes.len();
-                pb.set_message(&bytesize::to_string(count as u64, true));
+                pb.set_message(bytesize::to_string(count as u64, true));
                 buffer.put(bytes);
                 Ok(StreamWriter::Uploading {
                     buf_size,