+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.5.0 [Unreleased]
 * removed ending `/` when listing buckets
-* bump libraries (tokio 3)
+* fixed sub-command `get` to download file in the current path
 
 ## 0.4.0 ⛸
 * New sub-command `get` with option `-h` to return HeadObject.